Move tool-related headers from chain-lib #704
