+++ programs/create-genesis/ee_genesis_serializer.hpp
@@ -1,13 +1,14 @@
 #pragma once
 #include <boost/filesystem/fstream.hpp>
 #include <cyberway/chaindb/common.hpp>
-#include <cyberway/genesis/ee_genesis_container.hpp>
+#include <eosio/event_engine_plugin/ee_genesis_container.hpp>
 #include <eosio/chain/abi_serializer.hpp>
 
 namespace cyberway { namespace genesis {
 
 using namespace chaindb;
 namespace bfs = boost::filesystem;
+using namespace eosio;
 
 FC_DECLARE_EXCEPTION(ee_genesis_exception, 10000000, "event engine genesis create exception");
